add doc's

+++ order/src/main/java/ru/ik87/microservices/demo_shop/order/rest/controller/OrderController.java
@@ -75,6 +75,9 @@ public class OrderController {
         if (order == null) {
             throw new OrderNotFoundException(order_id);
         }
+        if (order.getStatus() != OrderStatus.CONFIRMED) {
+            throw new OrderBadRequestException(order_id);
+        }
         repository.delete(order);
     }
 }
